finished NgRx for list loading and partially detail
