+++ online-shop/src/app/components/containers/products-details/products-details.component.ts
@@ -1,29 +1,34 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
+import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
+import { AppState } from 'src/app/modules/shared/state/app.reducers';
+import { loadProductRequestAction } from 'src/app/modules/shared/state/product/product.actions';
+import { getProduct, getProducts } from 'src/app/modules/shared/state/product/product.reducers';
 import { Product } from 'src/app/modules/shared/types/product.types';
 import { ProductService } from 'src/app/services/products.service';
 
 @Component({
   selector: 'app-products-details',
-  templateUrl: './products-details.component.html',
-  styleUrls: ['./products-details.component.scss']
+  templateUrl: './products-details.component.html'
 })
 export class ProductsDetailsComponent implements OnInit {
-  product$ !: Observable<Product>;
-
+  product$ !: Observable<Product | null>;
+  
   constructor(
     private route: ActivatedRoute,
-    private service: ProductService 
+    private service: ProductService,
+    private store$:Store<AppState>
   ) { }
 
   ngOnInit(): void {
-    this.product$=this.getProduct();
+    this.loadProduct();
   }
 
-  getProduct(): Observable<Product> {
+  loadProduct():void{
     const id = String(this.route.snapshot.paramMap.get('id'));
-    return this.service.getProductById(id);
+    this.store$.dispatch(loadProductRequestAction({id}));
+    this.product$=this.store$.select(getProduct);
   }
 
   deleteProduct(id: string): void {
